Some cleanup and fixes.
