+++ .env.sample
@@ -3,7 +3,6 @@ FLASK_DEBUG=1
 FLASK_RUN_PORT=8001
 FLASK_RUN_HOST=127.0.0.1
 ASSET_DIR=/mnt
-DOWNLOAD_DIR=${ASSET_DIR}/downloads
 STORAGE_DIR=/home/data/storage_test
-BUILD_DB=1
+BUILD_DB=0
 DEVELOPER_MODE=0
