+++ README.md
@@ -7,12 +7,14 @@ At the moment, the server is used to resolve AAPB GUIDs to local file paths, and
 
 ## Usage 
 
+
 ### Within CLAMS apps
 
-The server deployment address is stored as [a organization variable](https://github.com/organizations/clamsproject/settings/variables/actions). To use the server (and `baapb` scheme in MMIF document locations), set `BAAPB_RESOLVER_ADDRESS` environment variable to the deployment address, and install the client plugin. 
+The server deployment address is stored as [an organization variable](https://github.com/organizations/clamsproject/settings/variables/actions). To use the server (and `baapb` scheme in MMIF document locations), set the `BAAPB_RESOLVER_ADDRESS` environment variable to the deployment address, and install the client plugin. 
 
+<!--
 All `brandeis` tagged pre-built container images (available in https://github.com/orgs/clamsproject/packages) 
-
+-->
 
 ### Server API
 
@@ -174,11 +176,21 @@ Install all the python dependencies with `pip install -r requirements.txt`, and 
 * `FLASK_APP`: must be `api`
 * `FLASK_DEBUG`: set to `1` to enable debug mode, otherwise `0`
 * `FLASK_RUN_PORT`: port number to listen on
-* `FLASK_RUN_HOST`: hostname to listen
+* `FLASK_RUN_HOST`: hostname
 * `ASSET_DIR`: path to the directory on the server where the AAPB media files (assets) are stored
-* `DOWNLOAD_DIR`: currently not in use
 * `STORAGE_DIR`: the directory where MMIF files are stored
 * `BUILD_DB`: set to `1` to build the database from scratch, otherwise `0`
 * `DEVELOPER_MODE`: set to `1`  for developer mode, which adds some routes to the API
 
-Start the server with `flask run`.
+Before you start the server for the first time you should build the assets database.
+
+```bash
+flask create-db
+```
+
+To start the server do
+
+```bash
+flask run
+```
+
