+++ app.py
@@ -1,0 +1,7 @@
+from api import create_app
+
+app = create_app()
+
+if __name__ == '__main__':
+
+	app.run()
