+++ packages/README.md
@@ -1,10 +1,10 @@
-Temporary directory with packages needed by the inspector part of the MMIF Server.
+Temporary directory with packages needed by this code, in particular by the
+inspector. These potentially include:
 
-These include:
+- A newer version of mmif-python than stipulated in the project file. This is
+  usually when a newer version of the summarizer is needed that is not available
+  on PyPI.
 
-- A new version of mmif-python (sporting a version number that
-it in the end may not have) that includes the summarizer and that may take a while to be available on PyPI.
-
-- An initial version of the MMIF Inspector, which is also not available yet.
+- An initial version of the MMIF Inspector, which is also not available on PyPI.
 
 Once the above packages are available on PyPI then this entire directory should be deleted.
