--- prototype/config.yml
@@ -1,1 +1,0 @@
-storage_dir: C:\Users\eljor\desktop\fast_playground\
