--- prototype/test_rewind.json
@@ -1,12 +1,0 @@
-{
-  "pipeline": {
-    "swt-detection/unresolvable": {
-    },
-
-    "simple-timepoints-stitcher/v1.3": {
-      "labelMap": "['B:bars', 'S:slate', 'S-H:slate', 'S-C:slate', 'S-D:slate', 'S-G:slate', 'I:chyron', 'N:chyron', 'Y:chyron', 'C:credits']"
-    }
-
-  },
-  "guid": "cpb-aacip-0515ac167c0"
-}
