--- prototype/test_zero_guid.json
@@ -1,21 +1,0 @@
-{
-  "pipeline": {
-    "swt-detection/unresolvable": {
-    },
-
-    "simple-timepoints-stitcher/v1.3": {
-      "labelMap": "['B:bars', 'S:slate', 'S-H:slate', 'S-C:slate', 'S-D:slate', 'S-G:slate', 'I:chyron', 'N:chyron', 'Y:chyron', 'C:credits']"
-    },
-
-    "doctr-wrapper/unresolvable": {
-      "tfLabel": "chyron",
-      "pretty": ""
-    },
-
-    "doctr-wrapper": {
-      "tfLabel": "['credits', 'credit']",
-      "pretty": ""
-    }
-
-  }
-}
